+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Features
 - Disable dashboards assistant chatbot if investigation feature flag enabled ([#626](https://github.com/opensearch-project/dashboards-assistant/pull/626))
 
+- detect serverless data source ([#625](https://github.com/opensearch-project/dashboards-assistant/pull/625))
+
 ### Enhancements
 
 ### Bug Fixes